oetree: add script
as a replacement for dot2tree.
The new script is hundreds of times faster and much more flexible.

It includes now package size information and render tweaks,
so even larger tree can be printed in a still readable fashion.

The previously mandatory endnode is now optional, also
the order of the inputs was adjusted to have a more common
look and feel across the different modes

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>